fix(importer): avoid raising an exception when handling invalid data
This commit makes record ID inference more exception-proof by handling
all exceptions. Because this function is already being called by the
exception handler for a previous failed attempt parsing a record, it is
guaranteed to struggle to successfully parse the record under a variety
of unknown circumstances, and so raising another exception is not
desired.