refactor: move workers and indexer out of `docker/` into `gcp/` (#2938)
#1714 
Move the indexer to `gcp/indexer` and the python workers into
`gcp/workers/`, so our deployed code is more together and separated from
our ci stuff.

I also want to deal with the shared worker Dockerfile & poetry
environment, but this PR already involved a lot of potentially risky
renaming so I've started with this.