fix(nvd-mirror): remove limit on cve references field (#3044)
Even though the NVD CVE API v2 schema says there's a maximum of 500
references, the response given for
[CVE-2009-3555](https://nvd.nist.gov/vuln/detail/cve-2009-3555) has 620.
This was causing a fatal error decoding the JSON response, stopping
`nvd-mirror` from completing.

I've just manually commented-out the auto-generated length check to
remediate this - not sure what the better long-term solution is for
this.