feat(api): switch Debian version query to range matching (#3047)
The current Debian's vulnerability matching only checking the affected
version. This is a problem for container scanning, as we only enumerate
versions during the import from source. So, many vulnerabilities are
mismatched if the package is slightly newer than the import time. For
example, the `linux` package at version 6.1.112-1 shows about 490
vulnerabilities for Debian 12. but at version 6.1.119-1, it only finds
14 vulnerabilities.

Switching to range matching to check both the affected range and
affected versions.