fix: failing vulnfeeds test (#3084)
This test started failing recently.
I *think* it's because the tag it was checking against was deleted from
the repo, so I've changed it to point to a new, existing tag & commit.

It now passes the tests, but I'm not sure if the test is still testing
what it's intending to.