refactor(terraform): read from the environment-specific configuration
This commit acts on reviewer feedback and refactors the previous work to
read from the environment-specific `kustomize.yaml`, rather than the
base CronJob YAML files. This means it picks up any environment-specific
CronJobs, notably `staging-api-test` (which is only present in staging).

It also DRYs things out slightly and reads the project_id from the `osv`
module, avoiding the need to declare it twice.