fix(debian-copyright-mirror): overhaul for more efficient run length
The k8s CronJob monitoring has highlighted that the
debian-copyright-mirror job appears to be overrunning. The wget
mirroring approach is grossly inefficient.

This replaces that with a more efficient parallel curl execution, using
a YAML file conveniently available to construct the URLs to download.

It sacrifices the attempt at only updating changed files, as this isn't
easily possible with curl.

The runtime is now < 10 minutes in local testing!