feat(vulnfeeds): better validate repos found in references
This commit tightens up the repo validation to exclude repos that don't
have any usable tags, unless the reference URL is obviously a commit.

The primary benefit here is that useless repos don't cause CVEs to be
in-scope for attempting conversion that shouldn't, while not
disregarding CVEs that have a potential fix commit as a reference, but a
repo that is otherwise useless for version-to-commit mapping.

In other words, it improves the overall conversion metrics by firming up
the denominator some more.