feat(exporter): add a safe_upload_single function (#3199)
This adds a file size delta function to prevent the unintended
overwriting of a file with a significantly smaller one.

It only uses it for the unified all.zip upload at present, as I'm not
sure how to identify when to use it for the ecosystem-specific zip files
as they seem to be bundled in with the individual record files.