feat: added a way to locally test frontend changes without connecting to GCP (#3229)
@michaelkedar and I created figured out how to utilize the datastore
emulator to mock the data from locally rather than connecting to
staging, so that small frontend tests, like checking behaviour
consequences of bad data, can be mocked without introducing bad data in
staging.

Future PRs could introduce a better data set and method: for example, a
public bucket or being able to upload a zip of OSV vulns to test on.

For now this suits the purpose of testing behaviour when there are
cycles in the graph.