feat(worker): include vulnerability ID in more logs (#3230)
This commit includes the record's vulnerability ID in a structured log
field for the worker's code path that imports updates to vulnerability
records.

After reviewing the current logger behaviour, it was determined that the
opportunity exists to add this existing structured log field in more
places, to aid with tracing the path of a specific record through the
worker.

This is intended to enable filtering logs emitted from `_do_update()`
for a specific vulnerability ID by using `jsonPayload.bug_id`