fix(api): Add CORS support to API (#3222)
The OSV API does not respond properly to CORS requests, for example a
call to https://api.osv.dev/v1/query with Origin in the headers should
return Access-Control-Allow-Origin: "*".

According to
https://cloud.google.com/endpoints/docs/openapi/specify-esp-v2-startup-options#cors
, this change to the ESPV2 env variables should get the job done without
requiring someone to go into the backend and mess with nginx / apache
files.

Why it's important: Integrations that use a webkit-based backend (or a
Web Application Framework) when configured to security best practices
will have CORS checks enabled for any non-same origin calls (meaning any
curl/fetch calls made outside of the domain of the tool). While this can
be bypassed by altering the Cross-Origin Resource Policy, it is not
recommended to do so as CORS prevents certain kinds of XSS/remote
hijacking attacks.

That means any integration system that does CORS checks cannot access
the OSV API.

There is precedent for enabling this within the integrations OSV already
is using - all of the CDN-pulled libraries and many of the data sources
such as NIST are correctly sending CORS headers when resources are
fetched from them. All NIST API endpoints have
Access-Control-Allow-Origin: "*" in the header. It is a well-documented
standard practice to enable this setting for public APIs, which OSV is.

Co-authored-by: Rex P <106129829+another-rex@users.noreply.github.com>