feat(exporter): adjust GCS client to reduce HTTP connection exhaustion (#3248)
Expand the HTTP connection pool, retries and blocking behaviour on
exhaustion to cater for the concurrent object uploads.

It appears that exporter runs are hanging for extended periods of time
due to connect exhaustion, based on investigation of a recent run:
https://cloudlogging.app.goo.gl/KxqkxJTG9ipNj46o8

This tries the technique described in
https://github.com/googleapis/python-storage/issues/253 to increase the
resources.

Also fix a linter complaint about the recently added
`safe_upload_single()`