feat(source): Add V8 to prod. (#3209)
Following up on #3205, we are adding V8 to the prod environment.

As per internal discussions, we are adding a V8 advisory to OSV.
The Chromium team has a
[policy](https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/security/updates.md)
that states that direct embedders of Chromium packages such as V8 are
responsible for using an up-to-date version of V8 packages. The
definition of that is that the V8 package is at most one week old.

Since these types of packages do not have their own identifiers and
therefore no advisory could be created, enforcement of such policies was
not possible. We now created a synthetic advisory generator that takes
such time-based policies and converts them into regularily updated
advisories that OSV can read:

https://github.com/google/chromium-policy-vulnfeed/blob/main/advisories/V8-advisory.json

---------

Signed-off-by: Viviana Sutedjo <viviana@sutedjo.net>
Co-authored-by: Oliver Chang <oliverchang@users.noreply.github.com>