refactor: Move upstream calculation from frontend to cron job (#3280)
We had some concerns about performance of the hierarchy calculation
being in the frontend so this should improve the frontend performance by
precalculating the hierarchy in the cron job instead. I would include
some metrics on how much performance is improved here, but the data
needs to be in staging to get this data.

Upstream hierarchy is stored in JSON format in 'reverse' order e.g. vuln
points at all its upstreams.

Upstream hierarchy must be calculated **AFTER** all transitive upstream
vulns in order to have the complete updated set of transitive upstreams
to calculate from.

Downstream hierarchy calculation will remain in the frontend as the
number of queries and time taken to calculate the hierarchy is
negligible, compared to the impact it would have on the cron job -
there's no way of checking for all of the vulns that exist as
downstreams so **every** `Bug` would need to be checked each time the
cron job was run, as opposed to the upstreams that we can check for the
presence of.

Other minor improvements include adding the 'expandible list' class to
the upstream and downstream frontend elements so that the list doesn't
overwhelm the page.

Relates to frontend for #3052