feat: Add go vanity metadata (#3300)
Add vanity URL metadata to return when go gets a package from osv.dev.

I think this is set up in a way where in the future we can also add
osv-scanner to the vanity URL by putting it on a sub path.

Needs to be merged before #3299 for it to work.