feat: Add osvdev pkg (#3299)
Add the osvdev package as a public package maanged in osvdev's
repository.

This also makes public the batch query helper to make it easier for
people to handle results with multiple pages.

While looking though this, I realized there's actually a bug with how
paging is being handled currently in osv-scanner, a typo here:


https://github.com/google/osv-scanner/blob/06755300671eaa9072327b505615462f29e2a879/internal/clients/clientimpl/osvmatcher/osvmatcher.go#L123-L127

means we only actually query two pages deep, and not any further. This
also fixes that issue. I'll make a separate PR in osv-scanner to fix it
there as well.