feat(frontend): Added weighted search results (#3356)
When a user searches for a specific vulnerability ID, that should be the
first record that shows up, not just the most recently updated record.

As such we added some extra ordering on record results: the ID given >
alias of the ID > in the ID's downstream > related to the ID.

May need to add functionality in the future for also including the ID's
upstreams. Unfortunately, this only works 100% as intended with one page
of results.