feat(frontend): Added rerouting for aliases and queries  (#3455)
Closes https://github.com/google/osv.dev/issues/760

Intended behaviour:

For an ID that we do not have in our datastore, that is part of an alias
group with 1 other vuln, using /{vuln-id} will redirect to the other
alias vuln. If there are more aliases in the group, or no alias at all,
the user will be rerouted to the search query page, with the chance of
finding a relevant downstream vuln.

Current best example is CVE-2024-40120. this may change if we import the
record at some point.

Also updated the emulator with some examples to test.