fix(staging-api-test): make load test resilient to connection errors (#3531)
The staging-api-test throws connection errors due to a race condition
where if a connection is closed prematurely, the client pool may reuse
it before the closure is recognized.

Changes it to wait for the full server response to fix this issue, and
also adds return_exceptions=True to the gather call to disable noise
from any other transient errors. We already log individual failures and
some are expected during a load test.