feat(website): Make determination of alias/related/upstream links faster (#3580)
Instead of sequentially querying the *whole* Bug in datastore to check
if it's in the database for rendering links in
alias/related/upstream/downstream fields, collect all the possible IDs
and query them beforehand. (This involved some slight refactoring to how
the up/downstream html is generated)

Locally, this made loading the page for
[USN-7449-1](https://osv.dev/vulnerability/USN-7449-1) go from >6
minutes to ~8 seconds (which is like a 50x speedup).
For reference, loading the live page on osv.dev took me just over 50
seconds without this improvement, so this should hopefully go down to
less than 1 second after it's merged.