fix(api): double backend concurrency (#3640)
Bump the max concurrent requests from 5 to 10 for the API backend, to
help with container instance scaling.
This is not for the batch query backend, which still is just 1.

Hopefully this won't have a notable performance impact. I don't really
want to change this and #3638 at the same time, since I want to be able
to measure the individual impacts of both. We could probably push this
out with the release this week, and let #3638 stage until next week.