perf(api): Improve asynchronous behaviour (#3638)
Change all(?) the ndb calls used throughout the API handlers to use the
async versions, and rewrote a few functions to not block on the
completion of futures that it doesn't need to (particularly, converting
the Bug to a proto, and populating the related/alias/upstream fields of
those protos).

Testing on my private instance, this should be a 3-5x speedup for
queries with lots of vulnerabilities.

I've also made the batch queries properly set the modified time from the
alias/upstream group (which might actually be a tiny performance loss)
to ensure the modified time is always the same for the batch query and
the get by ID.

I've added `@ndb.synctasklet` to the rpc handlers to let us use `yield`
instead of `.result()`, just to discourage usage of it (using
`.result()` in a tasklet can cause stack overflows)