refactor(vulnfeeds): Use OSVSchema Vulnerability structs instead of maintaining a duplicate set (#3714)
Noticed that we were using an outdated Vulnerability struct in vulns.go.
This will update vulnfeeds to use the osvschema go bindings instead,
which will keep the fields better up to date.

Also refactored FromCVE to take the specific fields it needs instead of
the whole CVE for better modularity.

Sorry for the long PR.