feat: Begin populating new database fields (#3708)
This PR begins our database table migrations!
This adds a few new entities into Datastore:
- `Vulnerability`: containing a small amount of information of a
vulnerability, with the (overall) modified date, as well as some raw
fields that are overwritten by our enrichment.
- `ListedVulnerability`: Minimal vulnerability information that is
needed for our website's `/list` page
- `AffectedVersions`: An index for API matching, which should help
simplify our API logic and improve performance.
Additionally, full vulnerabilities are stored in the GCS bucket, which
will eventually become the new source of truth for the vulnerability
records (deprecating the Bug entity)

To populate these I've added a new `_post_put_hook` on the Bug entity,
and added some logic to the alias & upstream cron jobs to update the
entities/GCS also.

I've only set this to run on the test instance (and when running tests
with the emulator) because I don't want this to block releases on prod
if it turns out this has significant performance impacts (especially
with the blocking writing to the bucket in the alias & upstream crons).

Some other misc things to note:
- I'm writing both `.pb` and `.json` files to to bucket, because the
proto files are a bit easier to work with in code, but the JSON is more
usable externally. It's probably better to just write the `.pb` and have
the exporter create the `.json` files
- I had to remove the transactions from the reput tools, because ndb
doesn't like nested transactions
- I have written my own (minimal) mocker for the `google.cloud.storage`
operations, and added that to the datastore emulator starter code
- I've made a few changes to some tests that were creating incomplete
Bugs (which caused issues when trying to convert them for writing)
- I have a TODO to create another worker/cron job that retries failed
GCS writes - that's probably the next step.
- Before we merge this, we should pause the alias/upstream cron jobs,
then re-put every Bug (which will take some time) to avoid noisy error
logs.