feat(frontend): Prevent horizontal overflow for affected packages (#3753)
This PR improves the UX of the "Affected packages" section on the
vulnerability page.
Previously, a large number of packages would [in some
cases](https://github.com/google/osv.dev/issues/3752) overflow the page
and become inaccessible.

This change modifies the rule for displaying the vertical layout to
include ecosystem names, ensuring all packages are viewable and
accessible.

This change also collapses the vertical tabs when they exceed a certain
threshold, to minimize vertical scrolling.

Fixes #3752 , related to #3087