fix: alpine versions `<=`/`>=` (#3978)
The alpine versions seem to be falling back to string comparisons for
`>=`/`<=`, which is wrong.
This issue is in the univers code we've vendored, which seems to be a
bit outdated, but it's [still present on their current
version](https://github.com/aboutcode-org/univers/issues/172)

I've added tests to catch this in Alpine and other ecosystems, but APK
seems to have been the only ecosystem affected.