feat(staging): record checker (#3957)
Adds a cron job to validate the consistency of records between GCS and
datastore with the new database format (#3850). This sends messages to
the recoverer (#3821) to attempt to repair.

The record-checker stores information from its previous runs in
datastore in a new JobData entity - where the datastore Key is the name
of the metadata, and the value is stored in a `value` property.

I've written this in Go to because it's fairly simple and to set a
precedent for writing / migrating other components away from python. I
made a copy of the `logging` submodule from `vulnfeeds` into the new
toplevel `go` directory - we'll want to consolidate these soon. There's
also room for refactoring the record checker code have some reusable
components for e.g. setting up datastore, gcs, or pub/sub.

I added a validation test script to make sure datastore entities written
in Python are compatible with the go definitions, and vice versa.