fix(api): page_tokens (#4089)
When package querying was paginated, it was always returning the 'not
started' page token, because you cannot have page tokens in 'OR'
queries. The exception was being caught and assumed to be the 'There is
no cursor currently' error that happens when you ask for a cursor before
iterating.

Removed the 'IN' query (so ecosystem-less queries will not match GIT
repos), and made the exception handler explicitly check for 'There is no
cursor currently' to help catch this in the future.