fix(api): temporary workaround for slow Ubuntu linux queries (#4091)
The huge number of Ubuntu linux vulns makes the new database method for
querying them take significantly longer when there are 0 results (e.g.
when querying the latest version).

While we think of how to fix this better, reimplemented the original way
these were being queried for Ubuntu linux vulns using the Bug entity and
*not* doing range-based matching.

(I *think* this is equivalent to the original code...)