feat(cve5): improve storing of conversion outcomes (#4150)
Started tracking counts of unresolved and resolved version metrics and
refactored where versions are extracted. We will be able to use this
later to do some analysis on successes, regressions, progression of
conversions later.

Moving the version extraction call allows us to make a base OSV record
using CVE5 data without extracting version data. This means we are able
to better control whether or not we actually want to attempt to extract
versioning information, for example if the cna is out of scope for
extraction.