fix: worker not updating converted record if override doesn't exist (#4164)
When checking for vulnerability overrides, if an override doesn't exist
for a specific vulnerability, the code proceeds without marshalling the
original vulnerability data. This means an empty buffer is passed to the
hashing function and all of our records attached to the upload worker
had the same empty input hash.

This is not ideal, as it means nothing gets updated...