feat: reimplement the exporter in Go (#4197)
More database migration!
Reimplement the exporter in Go, reading from the GCS proto files instead
of from the datastore Bugs.
I've made the whole thing more parallel and all in-memory, which should
be a pretty decent performance improvement.

Testing is currently missing - we need a way to mock the GCS buckets.