feat: introduce `apitester` tool (#4266)
Hopefully the readme explains most of this, but at a high-level this
introduces a new tool based off `osv-scanner` to make it easy to test
API changes.

Tests are provided as "cassettes" which we use `go-vcr` to load since
that has the bonus benefit of being [compatible with cassettes we'll be
using in
`osv-scanner`](https://github.com/google/osv-scanner/pull/2200), though
we do format the `request.body` to be multi-line so it's easier to
understand what the query under test is.

---------

Co-authored-by: Rex P <106129829+another-rex@users.noreply.github.com>