fix(bindings): use correct query parameter for DetermineVersion endpoint (#4314)
This PR fixes a bug in the `ExperimentalDetermineVersion` function.

Previously, the entire `api.DetermineVersionParameters` struct was
passed as the request body, however the API endpoint expects the nested
`Query` object instead. This change fixes it by passing `query.Query` to
the request, aligning it with the intended API usage.

The bug was not caught in the test because we don't have actual testing
for it...
https://github.com/google/osv.dev/blob/master/bindings/go/osvdev/osvdev_test.go#L422