fix(frontend): render backticks on summary fields (#4353)
There's an issue with backticks in the Overpass variable-weight font:
https://github.com/RedHatOfficial/Overpass/blob/master/fonts/variable/Overpass%5Bwght%5D.ttf
, which is the version hosted on Google Fonts and [used by the
frontend](https://github.com/google/osv.dev/issues/4345#issuecomment-3541453203)

This change adds a filter to escape summary text backticks and replaces
them with the monospace variant that does not have this issue. If the
Overpass font is fixed in upstream, this workaround should no longer be
needed.

Fixes #4345