feat: Have importer write Bugs before sending to worker (#4344)
Added a some logic to make the importer update the database when it
receives updates for the data sources, instead of having to wait for the
worker to finish enumerating versions / git commits.

To prevent churn, the importer checks if the affected packages have
changed in the upstream source first. If they haven't it won't overwrite
them (so we can keep our previous worker enrichment). The worker will
still run for every record, however.

Bugs put by the importer will use the modified date from the source,
while bugs modified by the worker now use the time the worker finished
(instead of the source time like before).

I've also modified the REST test to reduce an absolutely massive
vulnerability record, and made the `osv.utcnow()` return a time in the
future instead of in 2021, which was causing weirdness with time going
backwards.

This whole logic is a bit temporary - I eventually want to rewrite the
importer (in go) to stop relying on the Bug entities, and remove much of
the OSS-Fuzz-specific behaviour.