feat(CVE5): Add link to NVD and computed CVEList file and hash (#4422)
It only makes sense that we include the references to the places we
source our CVE data, both for triage, and just record keeping.

This PR also fixes a bug where empty urls are able to be saved, which
shouldn't happen based on the CVE5's quality requirements, but better
safe than sorry.

During this process, identifyPossibleURLs deduplication logic was moved
to its own function to be used better elsewhere if necessary.