feat: rewrite alias/upstream computation in Go (#4395)
This is pretty much a direct translation of the existing python code
(which makes some of it unidiomatic go), changed to use the new
`Vulnerability` instead of `Bug`.

One difference is the alias and upstream computations now run at the
same time, and the writes to the Vulnerability protos now happen
together. I had to modify the recoverer to be able to handle multiple
updating multiple fields when the GCS generation is mismatched to
support this properly.

Also ported over the datastore emulator manager, and made some
interfaces for the google cloud clients (with the added benefit of
introducing write retries which resolves #4355)
This has allowed my to bring over the existing tests.

I've renamed the `alias-computation` cron to `relations` - Cloud Deploy
is going to create a new workload for this, and the old one is going to
have to be deleted manually.

In a follow-up, I'm going to add another goroutine to calculate the
related field, since these aren't being kept up to date at the moment.