fix(website): allow non-double-encoded json upstreams on website (#4467)
The Python code seems to be assuming `upstream_hierarchy` is stored as a
JSON-encoded value within a JSON string.
e.g. `"{\"CVE-123\": [\"CVE-456\"]}"`. The new Go code just makes the
singly-encoded string (`{"CVE-123": ["CVE-456"]})`). (trying to generate
this old format in Go seems complicated)

Change the website to support both encodings to avoid throwing 500s when
one or the other is encountered